% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{countries_intersection}
\alias{countries_intersection}
\title{Extract country names from overlapping map polygons}
\usage{
countries_intersection(map)
}
\arguments{
\item{map}{An \code{sf} object of the species' spatial range.}
}
\value{
A character vector of country names that intersect the input map.
}
\description{
Identifies which country polygons intersect with a given species distribution map.
}
