% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{exportSpatialobject}
\alias{exportSpatialobject}
\title{Export a spatial object for IUCN Red List submission}
\usage{
exportSpatialobject(previous_map, new_map, currentyear)
}
\arguments{
\item{previous_map}{\code{sf} object from previous IUCN assessment.}

\item{new_map}{New \code{sf} geometry (processed or manually edited).}

\item{currentyear}{Numeric. The current assessment year.}
}
\value{
The exported geometry object.
}
\description{
Combines metadata and new geometry into an IUCN-compatible shapefile.
}
