% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{extractElevation}
\alias{extractElevation}
\title{Extract elevation for each occurrence record}
\usage{
extractElevation(occs)
}
\arguments{
\item{occs}{A data frame with decimal coordinates and occurrence information.}
}
\value{
The same data frame with a new column \code{elevationCalculated}.
}
\description{
Adds elevation values (in meters) to each record based on a global DEM raster.
}
