% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{manualEdition}
\alias{manualEdition}
\title{Manually edit a distribution map via interactive interface}
\usage{
manualEdition(modification_type, maps, previous_map, occs)
}
\arguments{
\item{modification_type}{Character. \code{"add"}, \code{"remove"}, or \code{"keep_as_drawn"}.}

\item{maps}{A list output from \code{checkMap()}, including two geometry layers.}

\item{previous_map}{The existing \code{sf} map to edit.}

\item{occs}{An \code{sf} object of valid occurrence points for reference.}
}
\value{
An \code{sf} object representing the manually modified geometry.
}
\description{
Allows users to add, remove, or redraw polygon areas using \code{mapview::editMap()}.
}
