% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{sRL_SimpleGBIF}
\alias{sRL_SimpleGBIF}
\title{Download GBIF occurrences for a species}
\usage{
sRL_SimpleGBIF(co_EXT)
}
\arguments{
\item{co_EXT}{A numeric vector of length 4 indicating longitude and latitude bounds (xmin, xmax, ymin, ymax).}
}
\value{
A data frame of GBIF records.
}
\description{
A simplified wrapper for downloading GBIF records within custom coordinate bounds.
}
